Fix spiral bot. Small improvements
 * Fix spiral bot: It works with the refactorings and also performs better :)
 * Pass along `playerId` to worker in INIT message.
 * Add a function `getMyCurve` to `Bot` to determine which curve belongs to the bot's player.
